refactor: 라우팅 수정
