+++ packages/mobile/src/page/Place/Report/style.module.scss
@@ -1,5 +1,5 @@
 @import "@shared/styles/color.scss";
-@import "../../../styles/main.scss";
+@import "../../../styles/main";
 
 .wrap {
   position: relative;
